+++ components/CustomCard/Profile/ProfileCard.js
@@ -1,3 +1,5 @@
+// 배효운
+
 import React, { useEffect, useState } from "react";
 import Card from "@mui/material/Card";
 import CardContent from "@mui/material/CardContent";
