+++ components/Grid/GridItem2.js
@@ -1,3 +1,5 @@
+// 배효운
+
 import React from "react";
 // nodejs library to set properties for components
 import PropTypes from "prop-types";
