+++ components/Professions/ProfessionsContainer.js
@@ -1,3 +1,5 @@
+// 배효운
+
 import React from "react";
 import Image from "next/image";
 import Box from "@mui/material/Box";
