+++ pages-sections/profile/profileSections/published/SectionCertificate.js
@@ -1,3 +1,5 @@
+// 배효운
+
 import * as React from "react";
 
 //components
