+++ pages/api/partner/index.js
@@ -1,3 +1,5 @@
+// 배효운
+
 import prisma from "../../../utilities/prisma/client";
 
 const handle = async (req, res) => {
