+++ pages/api/tags_resume/[type]/index.js
@@ -1,3 +1,5 @@
+// 배효운
+
 import { resolve } from "path";
 import prisma from "../../../../utilities/prisma/client";
