+++ pages/partner/index.js
@@ -1,3 +1,5 @@
+
+
 import { useEffect } from "react";
 import { useRouter } from "next/router";
