+++ pages/profile/[id]/index.js
@@ -1,3 +1,5 @@
+// 배효운
+
 import Link from "next/link";
 import { useRouter } from "next/router";
 import Card from "../../../components/CustomCard/Profile/ProfileCard";
