+++ pages/profile/create.js
@@ -1,3 +1,5 @@
+// 배효운
+
 import { useEffect, useState, useReducer } from "react";
 import { useRouter } from "next/router";
 //components
